Only check for pants.pex if test actually needs it
For example, Clippy never uses pants.pex, so it doesn't make sense to require it.

We keep the check, however, because it is a much nicer user experience to get the eager die() message that instructs users how to fix it, rather than trying to figure out on their own how to fix things.