Port `ci.sh` to Python 3 for more descriptive CLI flags and less duplication (#7849)
### Problem
`.travis.yml` can be cryptic to read due to `ci.sh`'s singular use of short CLI flags. For example, it is difficult to explain what this line does without checking the script or surrounding context:

```bash
./build-support/bin/ci.sh -fmrt2
```

### Solution
Port the script to modern Python 3, which allows us to use `argparse`.

Further, we can now better express choosing the Python version to use. Rather than having the flags `-2` and `-7`, now you specify `--python-version {2.7,3.6,3.7}` or `-v {2.7,3.6,3.7}`. You may only choose one, and it defaults to `3.6`.

Finally, using Python 3 allows us to improve the code itself, such as de-duplicating sections and avoiding materializing targets to disk with the `run_python_tests()` section.

### Result
CI will run the same, but will hopefully be easier to understand and to iterate on. For example, the above command becomes:

```bash
./build-support/bin/ci.py --githooks --sanity-checks --doc-gen --lint --python-version 2.7
```