Extract out `resolve_requirements` V2 rule for creating PEXes with requirements (#7846)
### Problem
`python_test_runner.py` does too much right now. 

One of its main pieces of logic that should be factored out is creating a PEX with the desired requirements, entry point, and interpreter constraints. This will presumably be useful for other V2 rules, so should not be coupled to the Pytest runner.

Also, it was found in https://github.com/pantsbuild/pants/issues/7795 that this requirements PEX is causing ~75% of the performance issues for the V2 Pytest runner. Having it extracted out into its own rule will make it easier to develop, test, and optimize.

### Solution
Create a generic rule, along with dedicated tests.