NailgunTasks execute java with consistent cwd.
Previously, nailgun-mode executions would run with the cwd set to the
buildroot but subprocess mode executions would inherit cwd. These
usually aligned since pants is run from the buildroot, but could
misalign in tests where a test buildroot is propped up.

Ensure cwd is always explicitly the buildroot for NailgunTasks and fix
executor cwd plumbing to effect this with a test added that proves this
all works.

Additionally, fix ScalaPlatform classpaths, which had been relativized
to the buildroot to support snapshotting. Maintain the snapshot relative
paths, but leave the in-memory classpath un-altered and populated with
absolute paths like all other runtime classpaths.

Prep work for #7866.