Fix Junit runner failing when `--extra-jvm-options` provided (#7877)
### Problem
In Python 3, you cannot compare unlike types like `list` and `tuple`, whereas Python 2 allowed this. https://github.com/pantsbuild/pants/pull/7067 mostly fixed this, but got the types wrong for `extra_jvm_options` and `extra_env_vars`.

This was detected when running Twitter code with Pants using Python 3, which resulted in the exception:

```python
Exception message: '<' not supported between instances of 'tuple' and 'list'
```

We did not encounter this error earlier because we never encountered the edge case of one target specifying `extra_jvm_options` and the other not specifying it, which would result in comparing a tuple to a list.

### Solution
For these two fields, there is no `None` value ever. The values are always tuples, just sometimes empty tuples. So, we can remove the fallback value.

### Result
The Twitter test now succeeds.