Fix Nailgun failure when the port is not specified (#7878)
Downstream Twitter tests were failing to compile certain targets with the error message:

```
zinc[zinc-java](path/to/target:target) failed: an integer is required (got type NoneType)
```

This happens when `ProcessManager.socket` returns `None`, which is an expected return value:

https://github.com/pantsbuild/pants/blob/e727747e036584301fb4dac5203597fd1066d6f7/src/python/pants/pantsd/process_manager.py#L306-L309

This ends up resulting in the nailgun code trying to use `None` for the `port`, which is invalid and causes the error.

Instead, we use Pants' idiom of setting the default to `None` and in the body of the constructor then defaulting to the actual default value we care about.