Use Selectors module for more robust and efficient syscalls (#7882)
### Problem
Twitter was running into the error `filedescriptor out of range in select()` when compiling a large number of targets, per https://github.com/pantsbuild/pants/issues/7880.

In the process, we found that it is more efficient and robust to use the more modern syscalls of `epoll()`, `devpoll()`, and `kqueue()` than `select()`. Python 3.4 introduced the Selectors module to choose the best syscall available for us.

This will close https://github.com/pantsbuild/pants/issues/7880.

### Solution
For Python 3 code, use the new selectors library. Python 2 code stays the same as before, since the library does not exist for it and we are soon going to drop Python 2.