Add Centos7 docker image to allow building Python 3.7 wheels (#7892)
Per #7288, we will very soon want to replace usage of Centos6 for Centos7 to pay off tech debt. The other motivation for using Centos7 is that we could start to build Python 3.7 wheels for Linux, meaning that we could ship a Python 3.7 PEX.

This is pre-work for both goals by adding the base Centos7 image, which we will publish to DockerHub.

See https://travis-ci.org/pantsbuild/pants/builds/547408642 for a sample run using this new Docker image.