Use S3 Sync for CI deploys to avoid recopying files (#7895)
We currently copy all of `dist/deploy` into S3, even if the file is an exact copy of something already in S3. This poses a real cost, both monetary and time.

Instead, here we use `aws s3 sync` to only transfer files that are different from what is currently in the bucket.

Will close https://github.com/pantsbuild/pants/issues/7258.