Add whitelist of folders that must remain Python 2 compatible (#7941)
Three of our folders may be run with Python 2, so they must keep their compatibility, e.g. `from __future__` imports.

If this impacted more code, we would want to create (restore) formal linters to enforce the compatibility. However, because the code is so small and CI will fail upon dropping Py2 support, we simply add comments to the files explaining this expectation and modify `check_headers.py` to allow these files to not use Py3 headers.

In the process, we refactor `check_headers.py` to use the `pathlib` module for much nicer path handling.