Make union error messages better
cf. #7819 , we should be printing out a more helpful error message in
the case that a `Get` request is unsatisfiable because of a type
mismatch involving union membership.

This commit defines two new FFI functions: one for getting a handle to
the underlying Python class representing a type, given a TypeId; and
another for checking whether or not a type is a union (by inspecting
the '_is_union' attribute). It also modifies the `Get` struct to track
the optional declared subject of a `Get` request.

Using these, if a Get request fails in `nodes.rs`, we look at that
declared subject (if it exists), and inspect the Optional type contained
within in order to see if it is a union. If so, we print a clearer error
message mentioning the name and description of the union. If not, we
fall back to the existing error message.

In order to correctly print the description of the union, we also needed
to alter the @union decorator to explicitly create an attr on
@union-decorated types called 'union_description', and search for this
attr with `project_str` in the new union error message handling code.