Union description
In order to correctly print the description of the union, alter the
@union decorator to explicitly create an attr on @union-decorated
types called 'union_description', and search for this attr
with `project_str` in the new union error message handling code. Also
make sure that `extern_get_handle_from_type_id` returns a handle
to the python class represented by a `TypeId`, not the `TypeId` python
object itself.