Use dedicated requirements.txt for `contrib.python.checks.checker` (#7982)
### Problem
`contrib.python.checks.checker` must keep support for Python 2, e.g. using `six`, which we will soon drop from the rest of the codebase.

Likewise, the module uses `pycodestyle` and `pyflakes`, neither of which are used anywhere else in the project.

We would like a way to provide requirements for specifically this module and to not pollute the rest of the project in doing so.

### Solution
Introduce a dedicated `3rdparty/requirements.txt` for the module, per https://www.pantsbuild.org/3rdparty_py.html#3rdpartypython.

We also indicate `pants_venv` should install these dependencies locally and `release.sh` should build 3rd party wheels for this so that the released PEX continues to work for the contrib package.