fix grpc threadpool size (#7974)
### Problem

The number of grpc completion queues on the host is tied to the requested remote concurrency. The remote concurrency numbers are likely to be much larger than the available cores on the host machine creating thread management overhead on the host.

### Solution

Use the built in GRPC EnvBuilder class which picks a sensible number of completion queue / threads based on the host proc count.

### Result

The user wont have to worry about strange side effects on the host when increasing remote concurrency.