Add type checking to several `util` files (#7998)
The util code—excluding the metaprogramming files—is a good place to start adding type hints to Pants because it is composed of simple pure functions and is used widely throughout the codebase. 

This adds some initial hints to act as a demo for type hints in Pants and get us slightly closer to critical mass.