relativize path for scalac classpath entry (#8011)
### Problem

When using remote hermetic zinc (without native-image) we were encountering an error where a path was not found during remote execution. Digging into it, it was obvious that the path being sent wasnt being rewritten/relativized.

### Solution

added relativization to all classpath entries for scalac, confirmed this fixes the issue locally.