Set up pants.remote.ini for remoting Python unit tests (#7990)
Per https://github.com/pantsbuild/pants/issues/7649, we are working to remote unit tests in CI.

This implements step #2, which allows us to run unit tests locally (several, but not all tests). Here, we set up `pants.remote.ini` so that users only must point to it and provide the authentication token.