[WIP] Move MyPy from isoltated goal into 'lint' goal
Registers MyPy as a 'lint' goal, and implements opt-in 'whitelist' system for type checking.
Targets to be type checked are tagged according to the '--whitelisted-tag-name' option, defaults to `type_checked`
as used from #7886. Add new mypy_tests for new logic, and update old tests and mypy instances to reflect lint goal change.

New mypy logic added to accomodate new opt-in strategy. Mypy will acts according to the following:
1. Filter out non-whitelisted target root
2. Check if all targets now in context are whitelisted, if not we throw an error
3. Continue as normal, collecting python source files and running mypy

See #7886 and #6742 for more context