Loosen constraints on the jvm heap size for tests in extra_jvm_options (#8106)
**Problem**
We set the default heap size to 1GB, and there was a test that trusted on that being a consistent number. However, the test doesn't need that to work. It's not testing the actual default heap size, but that we can modify the heap size if we pass an option.

This was causing travis failures: https://pastebin.com/knEMAWm7

**Solution**
Only output a number if the heap size is less than 100MB. We set it to be very small in the target that we are testing.

**Result**
Hopefully no failures.