Balance pytest test sharding (#8108)
### Problem

The pytest test sharding plugin uses mod of test index over the number of shards to balance across shards. That works fine when there are a lot more tests than shards, which used to ~always be the case... until we turned on `--no-fast` in travis, at which point each pytest invoke would observe only the tests for a single target. This meant that in many cases there were fewer tests than shards in an invoke.

The effect was that we biased toward overloading lower shard counts (because targets with less than a dozen test methods are very common).

### Solution

Balance across shards by hashing the pytest item "nodeid" (used elsewhere in this file for item display).

### Result

Fewer straggling shards in travis.