use jdeps to run dep-usage task (#8093)
### Problem
closes #7995 

The current implementation of the dep-usage goal uses analysis files which are output from zinc compiles to track dependencies between java and scala targets. We want to replace zinc with rsc eventually so we need to disconnect the dep usage and dep linting tasks from zinc output.
### Solution

We replace using zinc analysis files to track target dependencies with the the jdk tool `jdeps`. We also change the format for the product produced by the AnalysisExecution task from a nested `{target: {src_file: [dependencies]}}` to a single `{target: [dependencies]} which removes an intermediate layer of granularity, simplifying the shape of the analysis product. We do this because jdeps does not have any information about which source files compiled to which class files, and it cannot because that information is not contained in .class files.

### Result

Changes are transparent to the user with the exception of the format of one error message when the user requests pants lint with the `--jvm-dep-check-missing-direct-deps={warn,fatal}` previously the error message would reference the source file that had an implicit dependency, and what the dependency is, but now we just reference the target address which has the implicit dependency. This is still enough information to fix the linting issue, because dependencies are specified on a target level anyway.