Upgrade cloc to 1.80 and add error handling (#8111)
When running the external cloc process, allow it to raise an
exception if the process fails, to provide a better error message.

cf. https://github.com/AlDanial/cloc/issues/401, the current latest
version of cloc, version 1.82, is broken in such a way that it breaks
pants tests. Version 1.80 doesn't have this bug, so we can upgrade to
this version for now, pending a fix upstream.