Fail for missing classpath entries with hermetic execution. (#8037)
### Problem

When hermetic execution is used for `zinc` or `rsc` and `DirectoryDigests` are not available for input classpath entries, we currently trigger a warning. But this being only a warning (and not an error) definitely masks a few different potential bugs, and might allow more to slip in in the future.

### Solution

Convert the warning into an error, and fix all of the issues exposed by that change:
* In some `Task` configurations, `ResourcesTask` could run before the compilers, meaning that its resources needed to be digested. This also prepares for remote execution of tests.
* When `--no-use-classpath-jars` was in use:
  * `ClasspathEntry`s having digests was preventing them from having an equality match in order to be removed from the classpath.
  * `rsc` was universally adding the jar for a target to the classpath, even though it was empty.
  * Both the loose classpath and output jar needed to be captured in hermetic execution.
* Tools that were bootstrapped were not propagating the digests that had already been captured for them in #7835, which caused "global" scalac plugins to not have digests.

### Result

Hermetic execution has improved validation that should help to catch additional bugs in the future.