Mark float as natively encodable to silence hash_utils warning
Compiling with pants with the `-ldebug` flag used to produce a copious
amount of output with the following warning:
```
[DEBUG] pants.base.hash_utils:pid=42887: Our custom json encoder RunTrackerOptionEncoder is trying to hash a primitive type, but has gone throughchecking every other registered type class before. These checks are expensive,so you should consider registering the type <class 'float'> withinthis function (RunTrackerOptionEncoder.default)
```

Adding the type: `float` to the list of natively encodable types fixes
the issue.

It allegedly improves performance, and it makes the output of pants in
debug mode much more pallatable.