[pantsd] Clear the global state of subsystems earlier, plus refactorings (#8130)
**Problem**
In DaemonPantsRunner, we were clearing the global options state after we had warmed the v1 graph and ran every v2 rule. This means that those runs were running with potentially old state.

**Solution**
Move the clearing of Subsystem and related global state to before we compute v2 rules.
This state is re-initialized in LocalPantsRunner.parse_options(), right after clearing.

Commits should be independently reviewable. The first commit are some refactorings unrelated to this change, the second one is the actual change. I deemed this a small enough change to bunch them together.