Add test to check return of hex_16_digit_string function (#8133)
**Problem**
The existing test for the hex_16_digit_string function checks only the format and the length. It does not check if the input value passed to the function is used. It is important because the input value is a random value, the same should be the result of hex_16_digit_string function, thus it should use the input value.

**Result**
A test was added that checks that the result of the hex_16_digit_string function is consistent with the input value.