Reduce scope of Runtime.block_on
Assertions in the Future context end up panicking a background thread,
which means we don't necessarily flush stdout buffers, and the error
message we get is more along the lines of "Future cancelled" rather than
a useful assertion error.

Instead, only block on the Futuresy piece, and perform assertions on the
main test thread.