Deflake test (#8153)
The way backing in is implemented, we halve the backoff period per
successful request, rather than based on time. The `expect_only_good`
method makes an unpredictable number of requests, filling a time window
rather.

So this test would pass when the number of requests happened to align
with the time window (which appears to be common on travis's CPUs), and
fail if they got out of sync.

We currently have no tests for the actual backing in behaviour, but I
think that's mostly ok; we have tests that shows that we start using
servers again, and we can add more if we need to.