Create all_roots @rule for use in roots goal
The SourceRoots.all_roots function shouldn't be used in a v2 engine
rule, since it uses python standard library functionality to talk to the
filesystem driectly. This commit adds a new @rule `all_roots` which does
the same thing, but in a v2-engine-compatible way.

The basic strategy here is to add a method on the `SourceRoots` class
called `traverse` which gets a list of every source root path string
(with *-globs) that `SourceRoots`. `all_roots` uses this method to
construct `PathGlobs`, then creates `Snapshot`s from these to get a
collection of the paths that actually exist on the file system. It then
tests these against the trie contained within the `SourceRoots` instance
to pare them down to the output we actually want. `list_roots` can then
use the output from this rulelinstead of invoking the
`SourceRoots.all_roots`.

We can't yet get rid of `SourceRoots.all_roots` because go_buildgen.py
is still using it, but once it's replaced there it should be possible to
delete that code entirely.