Send Zipkin spans in json format in a separate process (#8150)
### Problem
In py_zipkin implementation, Zipkin spans are sent after they all are recorded.  When the number of spans is several hundred thousand the time overhead of encoding and sending spans is significant.
For example for 70000 spans is ~ 7 secs in thrift encoding and 16 seconds in JSON. 

Because how thrift encoding is implemented in py_zipkin the time to encode in json is smaller than in thrift.
Also, the overhead of sending that many numbers of spans is also too big.

### Solution
Zipkin spans are encoded in json and sent to the Zipkin server from a separate process.

The follow-up work can be:
- to make thrift encoding faster in py-zipkin.
- encode and send v2 spans from rust part (without passing them to python and encode and send them with py-zipkin)
- record fewer spans (e.g. do not record spans that take less than 100 milliseconds) 

### Result
This is a temporary solution that will allow to encode and send Zipkin spans without the noticeable overhead.