Don't check HTTP response size if gzip'd (#8116)
This check that the size in bytes of the downloaded file is the same as
the size given in the HTTP Content-Length header is broken when
downloading a compressed file. cf.
https://2.python-requests.org/en/master/user/quickstart/#response-content
the `iter_content` method is iterating over the *uncompressed* bytes of
the file, but the Content-Length header gives the *compressed* lengths
of the file. So we don't expect these two lengths to match when
downloading a compressed file, and shouldn't be testing that they do.