Pchevalier/local execution server (#8213)
## Problem
To test pants behaviour on remoting, the current practice consists of creating a ssh tunnel to one of the execution servers or running a local scoot server. This creates a barrier to quickly testing one's changes.

## Solution
Create a local_execution_server executable that leverages the mock execution server library.

## Result
This allows us to run an execution server on localhost that will expect exactly one request (described through CLI arguments) and react to it with a default operation.
We may want to enrich what this local_execution_server can do in the future (such as handle multiple requests), but only do the minimum needed at the moment. The CLI application can always be extended later on a per-need basis.