Fix the reporting remote integration test
The main issue that caused non-reproducible failures with this test was
the fact that we ran pants with the default speculation strategy, which
meant that we could get successful runs despite the mock servers not
doing their job if the local run was fast enough.

It turns out that my local testing was always successful, but travis
runs weren't.

To fix it,
* Pass the process-execution-speculation-strategy=none flag to pant.

  This allowed reliably getting a failure locally as the mock execution
  server's behaviour was too simple for actually satsifying pants'
  demands.

* Make the mock execution server return a response pretending that all
  the output files are empty (have the empty digest)
  The output file paths are given as a CLI argument to the server

* Pass the paths to the files produced by cloc ("report" and "ignored")
  in the pants invocation

Also clean up some print-debugging statements. As they were introduced
in the same PR, they will be squashed before merging, so I'm not
bothering to clean-up the git history wrt. them.