Use `./pants.pex`, not `./pants`, to run internal integration tests (#8183)
## Changes to get `--chroot` working
As prework to running integration tests with `--chroot` and then [remoting](https://github.com/pantsbuild/pants/issues/8113), we first must make two major changes to how we handle integration tests:

1) Having integration tests depend on `//:build_root`. This ensures that `get_buildroot()` properly resolves to the chrooted folder in `.pants.d`, rather than traversing all the way up to the original build root.
2) Using `./pants.pex` rather than `./pants` in integration tests. Because V1 strips the prefix of source roots, whereby `src/python/pants` becomes `pants`, we cannot also have a file named `pants` in the same build root. Instead, we can safely depend on `./pants.pex`.

## Autogeneration of `pants.pex`
Now that we depend on `pants.pex`, it is very easy for this file to fall out-of-date and to unintentionally be testing old code.

We work around this via the design in https://github.com/pantsbuild/pants/issues/8209. First, we start caching built PEXes in `~/.cache/pants`, with the ID derived from all the source files used for Pants. This allows us to quickly switch between branches without having to regenerate the PEX every time.

Then, we teach `./pants` to run `bootstrap_pants_pex.sh` whenever `test` is one of the goals and it's not CI. This will ensure that the file is always up-to-date.