Properly depend on plugins in chrooted integration tests (#8257)
### Problem

Most integration tests end up with `pants.ini` in their build root due to the dependency in `pants_test/option/util/fakes.py`:

https://github.com/pantsbuild/pants/blob/44128e8c9a10d20ed5988d08188aa51354a79568/tests/python/pants_test/option/util/BUILD#L4-L10

When this happens, `./pants.pex` will try to apply all the options in `pants.ini`, including loading all of the backend packages:

https://github.com/pantsbuild/pants/blob/44128e8c9a10d20ed5988d08188aa51354a79568/pants.ini#L47-L69

Most of these backend packages had no explicit dependencies declared via BUILD entries, so when most ITs were run with `--chroot` or `--v2`, they would fail to load the backend packages.

### Solution

Explicitly declare all plugin requirements as a part of the `pants_local_binary` target, which is what's used to bootstrap `pants.pex`.

Note that these must be declared as a part of `pants_local_binary`, rather than simply in the target `:int-test`, because the generated file needs to have all of the code necessary to run properly directly in the PEX file.

### Result

Further progress on https://github.com/pantsbuild/pants/issues/8113. Now the main remaining issue for remoting the majority of integration tests appears to be making the implicit dependencies on `examples` and `testprojects` explicit, which this PR makes an initial example of how this may be done.